add Attribute and replace AttributeOutputPlugin by OutputPlugin::<Attribute<T>>
